+++ include/Tensor.hpp
@@ -7,12 +7,12 @@
 class Tensor {
 public:
     size_t size;
-    double* data;
-    double* grad;
+    float* data;
+    float* grad;
     std::vector<int> shape;
     std::vector<int> strides;
     std::vector<const Tensor*> prev;
-    std::function<void(const double*)> _backward;
+    std::function<void(const float*)> _backward;
     std::string _op;
 
     Device device = Device::CPU;
@@ -20,11 +20,11 @@ public:
     Tensor to(Device target_device) const;
 
     Tensor(const std::vector<int>& shape, bool isParam = false, Device device = Device::CPU);
-    Tensor(const std::vector<double> data, const std::vector<int> shape, const std::vector<int> strides, bool isParam= false);
+    Tensor(const std::vector<float> data, const std::vector<int> shape, const std::vector<int> strides, bool isParam= false);
     // View constructor
-    Tensor(double* data_ptr, double* grad_ptr, const std::vector<int>& shape, const std::vector<int>& strides, Device device); 
+    Tensor(float* data_ptr, float* grad_ptr, const std::vector<int>& shape, const std::vector<int>& strides, Device device); 
 
-    double& at(const std::vector<int>& indices);
+    float& at(const std::vector<int>& indices);
     Tensor broadcastTo(const std::vector<int>& targetShape) const;
     Tensor transpose() const;
 
@@ -36,6 +36,6 @@ public:
     Tensor operator+(const Tensor& other) const;
     Tensor operator*(const Tensor& other) const;
     Tensor operator-(const Tensor& other) const;
-    Tensor pow(const double exp) const;
+    Tensor pow(const float exp) const;
     Tensor sum() const;
 };
